Attempt to fix the issue of oversized images (#2)
Remove the unused img-small class. Add a new image class to the images displayed in the carousel. This class is used to set images to a height of 280px which will scale down when the width of the page is decreased.

This will break compatibility for IE users unfortunately. However, given it's tiny market share nowadays, this may be a worthy trade-off