Add installation instructions
There was zero instructions prior to this, and that kinda sucks for the
end user.

Add a fragment for installation instructions (and maybe other things.)
Use the page generator to automatically insert into each device page.

This can still be useful for other things such as warnings, other
useful instructions etc.