cards: replace description with codename
Codename is fairly important in the AOSP custom ROM world; that's the
name we set for each device and what appears for the build zip file
name, and thus many users identify their device this way.

We aren't using the description section for each device anyway (what
more could we possibly describe about the device in question...) so
just make use of that as base for codename handling.