app: rename `Source` to `Kernel source`
GPL and all that. We can reuse this code to link to the kernel,
which *was* in the plan as seen in the ymls.

Adjust device ymls to work with it.

Also, remove broken handling of `source` in card header. It's supposed
to make the entire card a hyperlink, but this behaviour is broken. We
don't need it, anyway.